requirements.txt
alembic==0.9.5
autoenv==1.0.0
bcrypt==3.1.3
cffi==1.11.1
click==6.7
coverage==4.4.1
cryptography==1.7.2
flasgger==0.8.0
Flask==0.12.2
Flask-API==1.0
Flask-Bcrypt==0.7.1
Flask-Migrate==2.1.1
flask-paginate==0.5.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.0
gunicorn==19.7.1
idna==2.6
itsdangerous==0.24
Jinja2==2.9.6
jsonschema==2.6.0
Mako==1.0.7
MarkupSafe==1.0
mistune==0.7.4
nose==1.3.7
psycopg2==2.7.3.1
pyasn1==0.3.6
pycparser==2.18
PyJWT==1.5.3
python-dateutil==2.6.1
python-editor==1.0.3
PyYAML==3.12
six==1.11.0
SQLAlchemy==1.1.14
Werkzeug==0.12.2
